password != username validation for resets and package moves
