+++ open-health-manager-app/src/main/java/org/mitre/healthmanager/service/dto/DUAManagedUserVM.java
@@ -1,6 +1,4 @@
-package org.mitre.healthmanager.web.rest.vm;
-
-import org.mitre.healthmanager.service.dto.UserDUADTO;
+package org.mitre.healthmanager.service.dto;
 
 /**
  * View Model extending the ManagedUserVM class with a UserDUADTO object included
