+++ open-health-manager-app/src/main/java/org/mitre/healthmanager/web/rest/AccountResource.java
@@ -11,14 +11,14 @@ import org.mitre.healthmanager.security.SecurityUtils;
 import org.mitre.healthmanager.service.MailService;
 import org.mitre.healthmanager.service.UserService;
 import org.mitre.healthmanager.service.dto.AdminUserDTO;
+import org.mitre.healthmanager.service.dto.DUAManagedUserVM;
+import org.mitre.healthmanager.service.dto.KeyAndPasswordVM;
 import org.mitre.healthmanager.service.dto.PasswordChangeDTO;
 import org.mitre.healthmanager.service.dto.UserDUADTO;
 import org.mitre.healthmanager.web.rest.errors.BadRequestAlertException;
 import org.mitre.healthmanager.web.rest.errors.EmailAlreadyUsedException;
 import org.mitre.healthmanager.web.rest.errors.InvalidPasswordException;
 import org.mitre.healthmanager.web.rest.errors.LoginAlreadyUsedException;
-import org.mitre.healthmanager.web.rest.vm.DUAManagedUserVM;
-import org.mitre.healthmanager.web.rest.vm.KeyAndPasswordVM;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
