+++ open-health-manager-app/src/test/java/org/mitre/healthmanager/web/rest/UserResourceIT.java
@@ -25,8 +25,8 @@ import org.mitre.healthmanager.repository.FHIRPatientRepository;
 import org.mitre.healthmanager.repository.UserRepository;
 import org.mitre.healthmanager.security.AuthoritiesConstants;
 import org.mitre.healthmanager.service.dto.AdminUserDTO;
+import org.mitre.healthmanager.service.dto.ManagedUserVM;
 import org.mitre.healthmanager.service.mapper.UserMapper;
-import org.mitre.healthmanager.web.rest.vm.ManagedUserVM;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.mitre.healthmanager.service.UserService;
